Template Final Report
This template includes a model with all interactions, one with refined interactions (must be checked) and one with all correlations including reifined interactions. All must be checked. One beta (age) has been left out of the second model, because it was shit in the third. We argue that adding it as a between-subject interacting variable is better